move base Model class into base.py
