+++ peepingtom/core/base.py
@@ -150,3 +150,16 @@ class DataCrate(list):
 
     def __repr__(self):
         return f'<DataCrate{[db for db in self]}>'
+
+
+class Model(ABC):
+    """
+    A Model is an object from which a set of Particles can be derived
+    """
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+
+    @abstractmethod
+    def derive_particles(self):
+        pass
